tests/krate/publish: Extract `keywords` module
